refactor: import adapter modules by their own paths
Delete the eight adapter barrels under `browser/` and `server/` and point
every consumer at the implementation file. Tests that mocked a barrel are
split into one `vi.mock` per module.
