--- src/adapters/browser/cloudinary/index.ts
@@ -1,3 +1,0 @@
-import "client-only";
-
-export * from "./widget";
