--- src/adapters/browser/daum/index.ts
@@ -1,3 +1,0 @@
-import "client-only";
-
-export * from "./useDaumPopup";
