--- src/adapters/browser/deeplink/index.ts
@@ -1,3 +1,0 @@
-import "client-only";
-
-export * from "./open-app";
