--- src/adapters/browser/kakao/index.ts
@@ -1,3 +1,0 @@
-import "client-only";
-
-export * from "./useKakaoLoader";
