--- src/adapters/server/cookies/index.ts
@@ -1,5 +1,0 @@
-import "server-only";
-export * from "./get";
-export * from "./set";
-export * from "./delete";
-export type * from "./type";
