--- src/adapters/server/jose/index.ts
@@ -1,3 +1,0 @@
-import "server-only";
-export * from "./encrypt";
-export * from "./decrypt";
