--- src/adapters/server/nodemailer/index.ts
@@ -1,2 +1,0 @@
-import "server-only";
-export * from "./send";
