+++ src/services/auth.ts
@@ -2,9 +2,12 @@ import "server-only";
 import type { UserRole } from "@/core/domain";
 import { UserModel } from "@/models/user.model";
 import { dbConnect } from "@/db/connect";
-import { getCookie, deleteCookie, setCookie } from "@/adapters/server/cookies";
-import { decrypt, encrypt } from "@/adapters/server/jose";
-import { comparePasswords } from "@/adapters/server/bcrypt";
+import { getCookie } from "@/adapters/server/cookies/get";
+import { deleteCookie } from "@/adapters/server/cookies/delete";
+import { setCookie } from "@/adapters/server/cookies/set";
+import { decrypt } from "@/adapters/server/jose/decrypt";
+import { encrypt } from "@/adapters/server/jose/encrypt";
+import { comparePasswords } from "@/adapters/server/bcrypt/hash";
 import mongoose from "mongoose";
 import { AppError } from "@/core/domain";
 import type { AuthSession } from "@/core/schemas";
