+++ src/ui/components/organisms/ImageField.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { Upload, Plus } from "lucide-react";
-import { CloudinaryWidget } from "@/adapters/browser/cloudinary";
+import { CloudinaryWidget } from "@/adapters/browser/cloudinary/widget";
 import { ImagePreviewItem } from "@/ui/components/organisms";
 import { Button, TypographyMuted } from "@/ui/components/atoms";
